Taking apart array data
I added three php files. The first presidents uses a longer method involving a foreach loop that uses a fat arrow. The file displays the presidents and their years in office. The second presidents file displays how to take data that uses tokens and takes it apart and displays it. We use the same method in the passwordFields file but take the data and use it to display different message for optional or extra fields.